+++ src/main/java/com/iexec/blockchain/command/task/initialize/TaskInitializeBlockchainService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,8 @@
 
 package com.iexec.blockchain.command.task.initialize;
 
-
+import com.iexec.blockchain.chain.IexecHubService;
 import com.iexec.blockchain.command.generic.CommandBlockchain;
-import com.iexec.blockchain.tool.IexecHubService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;