+++ src/test/java/com/iexec/blockchain/config/PublicConfigurationControllerTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2023-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package com.iexec.blockchain.config;
 
-import com.iexec.blockchain.tool.ChainConfig;
+import com.iexec.blockchain.chain.ChainConfig;
 import com.iexec.common.config.PublicChainConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;